pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[project]
name = "spotify-server"
version = "0.1.0"
description = "Kurzbeschreibung deines Projekts"
authors = [{ name = "Linus Meinders", email = "linus.meinders@web.de" }]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "spotipy", 
    "flask", 
    "rapidfuzz", 
    "pymysql", 
    "python-dotenv",
    "flask-sqlalchemy"
]

[project.scripts]
spotify-server = "spotify_server.run:main"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests"]

[tool.setuptools_scm]